New Thurster has consumed its brethren... it alone stands on the top of a mound of corpses
