+++ src/controller/config/thruster_config.yaml
@@ -1,72 +1,44 @@
+# Position and direction of each thruster on the ROV
+# - Direction is a unit vector indicating the force vector applied to the ROV when the motor is active
+# - Position is the location of the thruster relative to the ROV's center of mass, in meters
+
 thrusters:
-  # Forward left lateral thruster
   - id: 1
-    name: "front_left_lateral"
-    max_duty: 1.0             # max output from -1 to 1
-    angle_deg: -45             # thruster mounted 45° from forward axis
-    position: "lateral"
-    notes: "Used for forward-left and yaw control"
-    lever_arm_distance: 0.2   # metres
+    name: front_left_lateral
+    position_m: [-0.2, 0.2, 0.0]
+    direction: [0.7071, 0.7071, 0.0]
 
-  # Forward right lateral thruster
   - id: 2
-    name: "front_right_lateral"
-    max_duty: 1.0
-    angle_deg: 45
-    position: "lateral"
-    notes: "Used for forward-right and yaw control"
-    lever_arm_distance: 0.2   # metres
+    name: front_right_lateral
+    position_m: [0.2, 0.2, 0.0]
+    direction: [-0.7071, 0.7071, 0.0]
 
-  # Rear left lateral thruster
   - id: 3
-    name: "rear_left_lateral"
-    max_duty: 1.0
-    angle_deg: -135
-    position: "lateral"
-    notes: "Used for backward-left and yaw control"
-    lever_arm_distance: 0.2   # metres
+    name: rear_left_lateral
+    position_m: [-0.2, -0.2, 0.0]
+    direction: [-0.7071, 0.7071, 0.0]
 
-  # Rear right lateral thruster
   - id: 4
-    name: "rear_right_lateral"
-    max_duty: 1.0
-    angle_deg: 135
-    position: "lateral"
-    notes: "Used for backward-right and yaw control"
-    lever_arm_distance: 0.2   # metres
+    name: rear_right_lateral
+    position_m: [0.2, -0.2, 0.0]
+    direction: [0.7071, 0.7071, 0.0]
 
-  # Forward left vertical thruster
   - id: 5
-    name: "forward_left_vertical"
-    max_duty: 1.0          
-    angle_deg: 45         
-    position: "vertical"      
-    notes: "Used for heave up"
-    lever_arm_distance: 0.2   # metres
+    name: forward_left_vertical
+    position_m: [-0.2, 0.2, 0.0]
+    direction: [0.0, 0.0, 1.0]
 
-  # Forward right vertical thruster
   - id: 6
-    name: "forward_right_vertical"
-    max_duty: 1.0
-    angle_deg: -45
-    position: "vertical"
-    notes: "Used for heave up"
-    lever_arm_distance: 0.2   # metres
+    name: forward_right_vertical
+    position_m: [0.2, 0.2, 0.0]
+    direction: [0.0, 0.0, 1.0]
 
-  # Rear left vertical thruster
   - id: 7
-    name: "rear_left_vertical"
-    max_duty: 1.0
-    angle_deg: 135
-    position: "vertical"
-    notes: "Used for heave down"
-    lever_arm_distance: 0.2   # metres
+    name: rear_left_vertical
+    position_m: [-0.2, -0.2, 0.0]
+    direction: [0.0, 0.0, 1.0]
 
-  # Rear right vertical thruster
   - id: 8
-    name: "rear_right_vertical"
-    max_duty: 1.0
-    angle_deg: -135
-    position: "vertical"
-    notes: "Used for heave down"
-    lever_arm_distance: 0.2   # metres
+    name: rear_right_vertical
+    position_m: [0.2, -0.2, 0.0]
+    direction: [0.0, 0.0, 1.0]
