+++ src/controller/controller/thruster_controller_node.py
@@ -1,6 +1,7 @@
 import rclpy
 from rclpy.node import Node
-from sensor_msgs.msg import Joy
+from geometry_msgs.msg import Twist
+from std_msgs import msg
 from std_msgs.msg import Float32
 from ament_index_python.packages import get_package_share_directory
 import yaml
@@ -9,177 +10,147 @@ import numpy as np
 
 class ThrusterController(Node):
     def __init__(self):
+        # ROS2 stuff
         super().__init__('thruster_controller')
 
-        pkg_share = get_package_share_directory('controller')
-        # === Parameters ===
-        self.declare_parameter('joy_config', os.path.join(pkg_share, 'config', 'joystick_config.yaml'))
-        self.declare_parameter('thruster_config', os.path.join(pkg_share, 'config', 'thruster_config.yaml'))
-        self.declare_parameter('joy_topic', '/joy')
-        self.declare_parameter('thruster_topic', '/thruster')
+        # Parameters
 
+        # Four thrusters control linear velocity, and they are pointed at 45 degrees. To get an accurate force to pwm use the equation below
+        thruster_max = 4 / np.sqrt(2) 
+        self.declare_parameter('max_x_n', thruster_max)
+        self.declare_parameter('max_y_n', thruster_max)
+        self.declare_parameter('max_z_n', 4.0)
+        self.declare_parameter('max_roll_nm', 1.0)
+        self.declare_parameter('max_pitch_nm', 1.0)
+        self.declare_parameter('max_yaw_nm', 1.0)
 
-        joy_config_path = self.get_parameter('joy_config').value
-        thruster_config_path = self.get_parameter('thruster_config').value
-        joy_topic = self.get_parameter('joy_topic').value
-        thruster_topic = self.get_parameter('thruster_topic').value
-
-        self.get_logger().info(f"Loading joystick config from: {joy_config_path}")
-        self.get_logger().info(f"Loading thruster config from: {thruster_config_path}")
-        # === Load configurations ===
-        self.joy_map = self.load_yaml(joy_config_path)
-        self.thruster_map = self.load_yaml(thruster_config_path)
-        self.get_logger().info(str(self.thruster_map))
-        # === Subscribers / Publishers ===
-        self.joy_sub = self.create_subscription(Joy, joy_topic, self.joy_callback, 10)
-        self.thruster_pubs = [self.create_publisher(Float32, thruster_topic+'/thruster_'+str(i), 10) for i in range(len(self.thruster_map['thrusters']))]
-        
-        # === Internal state ===
-        self.last_joy_msg = None
-        
-        # === Compute Thruster Allocation Matrix ===
-        self.A = self.compute_thruster_allocation_matrix()
-        self.A_pinv = np.linalg.pinv(self.A)  # Compute the pseudoinverse of A
-
-        self.get_logger().info("Thruster Controller Initialized")
-
-    def load_yaml(self, path):
-        """Load a YAML file from a relative or absolute path."""
-        if not os.path.exists(path):
-            self.get_logger().warn(f"Config file not found: {path}")
-            return {}
-        with open(path, 'r') as f:
-            return yaml.safe_load(f)
-
-    def joy_callback(self, msg: Joy):
-        """Handle joystick input."""
-        self.last_joy_msg = msg
-        axis_values, button_values = self.parse_joystick(msg)
-        self.get_logger().info(str(axis_values))
-        thruster_outputs = self.compute_thrusters(axis_values, button_values)
-
-        # Publish the result
-        self.publish_thrusters(thruster_outputs)
-
-    def parse_joystick(self, msg: Joy):
-        axis_values = {}
-        button_values = {}
-
-        cfg_list = self.joy_map.get("joystick", [])
-        if not isinstance(cfg_list, list):
-            return {}, {}  # config malformed
-
-        for cfg in cfg_list:
-            # --- Axis entry ---
-            if "axis" in cfg:
-                axis_name = cfg["axis"]
-                axis_index = int(cfg.get("input", 0))
-                invert = bool(cfg.get("invert", False))
-                sensitivity = float(cfg.get("sensitivity", 1.0))
-                scale = cfg.get("scale", "linear")
-
-                raw = msg.axes[axis_index] if axis_index < len(msg.axes) else 0.0
-                if invert:
-                    raw *= -1.0
-
-                val = raw * sensitivity
-
-                # Optional: support your "scale" field (keep simple)
-                if scale == "logarithmic":
-                    # compress near 0, preserve sign
-                    val = np.sign(val) * np.log1p(abs(val))
-
-                axis_values[axis_name] = float(val)
-                continue
-
-            # --- Button entry ---
-            if "button" in cfg:
-                btn_name = cfg["button"]
-                btn_index = int(cfg.get("input", 0))
-                invert = bool(cfg.get("invert", False))
-
-                raw = msg.buttons[btn_index] if btn_index < len(msg.buttons) else 0
-                pressed = (1 - raw) if invert else raw
-                button_values[btn_name] = int(pressed)
-                continue
-
-            # Otherwise: ignore unknown entry types
-
-        # If you want buttons too, return both (or store button_values on self)
-        # For now, just store on self so you can use it later:
-        self.last_buttons = button_values
-
-        return axis_values, button_values
-
-    def compute_thruster_allocation_matrix(self):
-        """
-        Builds the transformation matrix A that maps thruster outputs -> vehicle forces:
-        [strafe, drive_forward, yaw, pitch, roll, up]^T = A @ [T1..T8]^T
-        """
-        
-        thrusters = self.thruster_map.get("thrusters", [])
-        A = np.zeros((6, len(thrusters)))  # 6 DOFs × N thrusters
-        
-        # strafe, drive_forward, yaw, pitch, roll, up
-
-        for i, thruster in enumerate(thrusters):
-            angle_deg = thruster.get("angle_deg", 0.0)
-            angle_rad = np.deg2rad(angle_deg)
-            lever_arm = thruster.get("lever_arm_distance", 0.0)
-            pos = thruster.get("position", "").lower()
-
-            # Horizontal thrusters: contribute to pitch, roll, and yaw
-            if pos == "lateral":
-                A[1, i] = np.cos(angle_rad) # linear y
-                A[0, i] = np.sin(angle_rad) # linear x
-
-                # Yaw contribution (direction depends on front/rear and rotation sense)
-                if angle_deg == -45 or angle_deg == 135:
-                    A[2, i] = -1
-                else:
-                    A[2, i] = 1
-
-            # Vertical thrusters: contribute only to up
-            else:
-                A[5, i] = 1  # +1 for up, -1 for down
-                A[4, i] = -np.sign(angle_deg)
-                A[3, i] = -1 if (np.abs(angle_deg) < 90) else 1
-
-        self.get_logger().info(f"Thruster allocation matrix:\n{A}")
-        return A
-
-
-    def compute_thrusters(self, axis_values, button_values):
-        """
-        Map joystick input (pitch, roll, up, yaw) to 8 thruster commands.
-        """
-        
-        # strafe, drive_forward, yaw, pitch, roll, up
-
-        U = np.array([
-            axis_values.get("strafe", 0.0),
-            axis_values.get("drive_forward", 0.0),
-            axis_values.get("yaw", 0.0),
-            axis_values.get("pitch", 0.0),
-            button_values.get("roll_left", 0.0) - button_values.get("roll_right", 0.0),
-            0.5 * (axis_values.get("up", 0.0) - axis_values.get("down", 0.0))
+        self.wrench_scale = np.array([
+            float(self.get_parameter('max_x_n').value),
+            float(self.get_parameter('max_y_n').value),
+            float(self.get_parameter('max_z_n').value),
+            float(self.get_parameter('max_roll_nm').value),
+            float(self.get_parameter('max_pitch_nm').value),
+            float(self.get_parameter('max_yaw_nm').value),
         ])
 
-        T = self.A_pinv @ U  # Thruster outputs
+        self.declare_parameter('neutral_us', 1514.0)
+        self.declare_parameter('min_us', 1100.0)
+        self.declare_parameter('max_us', 1900.0)
+        self.declare_parameter('max_force_n', 1.0)     # force at max PWM per thruster
+        self.declare_parameter('publish_rate_hz', 50.0)
+        pkg_share = get_package_share_directory('controller')
+        self.declare_parameter(
+                'thruster_config', 
+                os.path.join(pkg_share, 'config', 'thruster_config.yaml')
+                )
+        self.declare_parameter('thruster_topic', '/thruster')
 
-        # Clamp to [-1, 1]
-        T = np.clip(T, -1.0, 1.0)
+        self.neutral_us = float(self.get_parameter('neutral_us').value or 1514.0)
+        self.min_us = float(self.get_parameter('min_us').value or 1100.0)
+        self.max_us = float(self.get_parameter('max_us').value or 1900.0)
+        self.max_force = float(self.get_parameter('max_force_n').value or 50.0)
+        rate = float(self.get_parameter('publish_rate_hz').value or 50.0)
 
-        return T.tolist()
+        thruster_config_path = self.get_parameter('thruster_config').value
+        thruster_topic = self.get_parameter('thruster_topic').value
 
-    def publish_thrusters(self, thruster_outputs):
-        """Publish the computed thruster duty cycles."""
-        # self.get_logger().info(str(thruster_outputs))
-        for i in range(len(self.thruster_pubs)):
+        self.get_logger().info(f"Loading thruster config from: {thruster_config_path}")
+
+        # Configure Thrusters and Allocation
+        config = self.load_yaml(thruster_config_path)
+        self.allocMatrix = self.compute_thruster_allocation_matrix(config)
+        self.n_thrusters = self.allocMatrix.shape[1]
+        self.allocMatrix_inverse = np.linalg.pinv(self.allocMatrix)
+
+        # Subscribers / Publishers
+        self.status_sub = self.create_subscription(Twist, "velocity_commands", self.Status_Callback, 10)
+        self.thruster_pubs = [] 
+        for indx in range(self.n_thrusters):
+            pub = self.create_publisher(Float32, f'thruster/thruster_{indx}', 10)
+            self.thruster_pubs.append(pub)
+
+        # State
+        self.lastest_twist = Twist()
+        self.pwm_commands = np.full(self.n_thrusters, self.neutral_us)
+        # Listen to state
+        # Scream to thrusters
+
+        self.create_timer(1.0 / rate, self.publish_thrusters)
+        self.get_logger().info("Thruster Controller Initialized")
+
+
+    def load_yaml(self, config_path) -> dict:
+        """Load a YAML file from a relative or absolute path."""
+        if not os.path.exists(config_path):
+            self.get_logger().warn(f"Config file not found: {config_path}")
+            return {}
+        with open(config_path, 'r') as f:
+            return yaml.safe_load(f)
+
+    def compute_thruster_allocation_matrix(self, config : dict) -> np.ndarray:
+
+        thrusters = config['thrusters']
+        AllocMatrix = np.zeros((6, len(thrusters)))
+
+        for indx, thruster in enumerate(thrusters):
+            pos_m = np.array(thruster['position_m'], dtype=float)
+            dir = np.array(thruster['direction'], dtype=float)
+
+            dir = dir / np.linalg.norm(dir) # normalize the direction vector
+
+            # Linear Force Contribution
+            AllocMatrix[0:3, indx] = dir
+            AllocMatrix[3:6, indx] = np.cross(pos_m, dir) 
+
+            
+        self.get_logger().info(f"Thruster Allocation Matrix:\n {AllocMatrix}")
+        return AllocMatrix
+
+    def Status_Callback(self, msg: Twist):
+        """
+        Process a new velocity vector and turns them into PWM signals for the Thrusters.
+        """
+        self.last_state_msg = msg
+        # Get twist msg
+        # Twist -> msg -> force -> PWM
+        # Convert to Numpy vectors, so easier to work with
+        linearVel = np.array([msg.linear.x, msg.linear.y, msg.linear.z])
+        angularVel = np.array([msg.angular.x, msg.angular.y, msg.angular.z])
+        
+        torqueList = self.map_twist_to_toque(linearVel, angularVel)
+        # Publish PWM data
+        self.pwm_commands = self.map_torque_to_PWM(torqueList)
+
+    def map_twist_to_toque(self, linear : np.ndarray, angular : np.ndarray) -> np.ndarray:
+        """
+        Turns a twist msg -> a torque map for each thruster
+        """
+        tau_norm = np.concatenate([linear, angular])
+        tau = tau_norm * self.wrench_scale
+        forces = self.allocMatrix_inverse @ tau
+        return forces
+    def map_torque_to_PWM(self, forces : np.ndarray) -> np.ndarray:
+        """
+        Turns a list of torques -> to a list of PWM signals
+        """
+
+        normalized = np.clip(forces / self.max_force, -1, 1)
+        
+        pwm = np.where(
+                normalized >= 0,
+                self.neutral_us + normalized * (self.max_us - self.neutral_us),
+                self.neutral_us + normalized * (self.neutral_us - self.min_us)
+                )
+        return pwm
+
+    def publish_thrusters(self):
+        """
+        Publish current PWM commands to each thruster
+        """
+        for indx, pub in enumerate(self.thruster_pubs):
             msg = Float32()
-            msg.data = thruster_outputs[i]
-            self.thruster_pubs[i].publish(msg)
-
+            msg.data = float(self.pwm_commands[indx])
+            pub.publish(msg)
 
 def main(args=None):
     rclpy.init(args=args)
