+++ src/thruster_pkg/thruster_pkg/thrusters_full_test.py
@@ -27,7 +27,7 @@ time.sleep = lambda *_: None
 
 import rclpy
 from rclpy.executors import MultiThreadedExecutor
-from thruster_pkg.new_thruster_driver import ThrusterNode
+from thruster_pkg.thruster_pkg.thruster_driver import ThrusterNode
 
 
 def main():
